+++ modules/product/src/main/java/com/opengamma/strata/product/index/type/ImmutableIborFutureConvention.java
@@ -125,7 +125,7 @@ public final class ImmutableIborFutureConvention
       LocalDate tradeDate,
       Period minimumPeriod,
       int sequenceNumber,
-      long quantity,
+      double quantity,
       double notional,
       double price,
       ReferenceData refData) {
