+++ modules/product/src/test/java/com/opengamma/strata/product/equity/EquityTradeTest.java
@@ -24,8 +24,8 @@ public class EquityTradeTest {
   private static final TradeInfo TRADE_INFO = TradeInfo.of(date(2016, 6, 30));
   private static final Equity PRODUCT = EquityTest.sut();
   private static final Equity PRODUCT2 = EquityTest.sut2();
-  private static final int QUANTITY = 100;
-  private static final int QUANTITY2 = 200;
+  private static final double QUANTITY = 100;
+  private static final double QUANTITY2 = 200;
   private static final double PRICE = 123.50;
   private static final double PRICE2 = 120.50;
